+++ src/test/java/com/github/ddth/kafka/qnd/QndConsumerManual.java
@@ -8,49 +8,27 @@ import com.github.ddth.kafka.KafkaMessage;
 public class QndConsumerManual {
 
     private final static int NUM_MSGS = 1000;
-    private final static String TOPIC = "demo1";
-    private final static String GROUP_ID = "mygroupid";
+    private final static String TOPIC = "ddth-kafka";
+    private final static String GROUP_ID = "ddth-kafka";
     private final String BOOTSTRAP_SERVERS = "localhost:9092";
 
-    public void qndAsyncProducer() throws Exception {
-        System.out.println("========== QND: Async Producer ==========");
-        final boolean CONSUME_FROM_BEGINNING = true;
-        final KafkaClient.ProducerType PRODUCER_TYPE = KafkaClient.ProducerType.FULL_ASYNC;
+    public void flush(KafkaClient kafkaClient, String groupId, String topic) {
+        kafkaClient.seekToEnd(groupId, groupId);
 
-        try (KafkaClient kafkaClient = new KafkaClient(BOOTSTRAP_SERVERS)) {
-            kafkaClient.init();
-
-            long timestart = System.currentTimeMillis();
-            long RECEIVED_MSGS = 0;
-            for (int i = 0; i < NUM_MSGS; i++) {
-                KafkaMessage msg = new KafkaMessage(TOPIC,
-                        "message - " + i + ": " + System.currentTimeMillis());
-                kafkaClient.sendMessage(PRODUCER_TYPE, msg);
-                long t1 = System.currentTimeMillis();
-                msg = kafkaClient.consumeMessage(GROUP_ID, CONSUME_FROM_BEGINNING, TOPIC, 10000,
-                        TimeUnit.MILLISECONDS);
-                long t2 = System.currentTimeMillis();
-                if (msg != null) {
-                    RECEIVED_MSGS++;
-                }
-                // System.out.println((msg != null ? msg.contentAsString() :
-                // null) + "\t" + (t2 - t1));
-            }
-            long timeEnd = System.currentTimeMillis();
-            long d = timeEnd - timestart;
-            System.out.println("Total: " + RECEIVED_MSGS + " msgs in " + d + "ms / "
-                    + (RECEIVED_MSGS * 1000.0 / (double) d) + " msg/sec");
-            Thread.sleep(2000);
+        KafkaMessage msg = kafkaClient.consumeMessage(groupId, topic);
+        while (msg != null) {
+            msg = kafkaClient.consumeMessage(groupId, topic);
         }
     }
 
     public void qndSyncNoAckProducer() throws Exception {
-        System.out.println("========== QND: SyncNoAck Producer ==========");
+        System.out.println("========== QND: NoAck Producer ==========");
         final boolean CONSUME_FROM_BEGINNING = true;
-        final KafkaClient.ProducerType PRODUCER_TYPE = KafkaClient.ProducerType.SYNC_NO_ACK;
+        final KafkaClient.ProducerType PRODUCER_TYPE = KafkaClient.ProducerType.NO_ACK;
 
         try (KafkaClient kafkaClient = new KafkaClient(BOOTSTRAP_SERVERS)) {
             kafkaClient.init();
+            flush(kafkaClient, GROUP_ID, TOPIC);
 
             long timestart = System.currentTimeMillis();
             long RECEIVED_MSGS = 0;
@@ -77,12 +55,13 @@ public class QndConsumerManual {
     }
 
     public void qndSyncLeaderAckProducer() throws Exception {
-        System.out.println("========== QND: SyncLeaderAck Producer ==========");
+        System.out.println("========== QND: LeaderAck Producer ==========");
         final boolean CONSUME_FROM_BEGINNING = true;
-        final KafkaClient.ProducerType PRODUCER_TYPE = KafkaClient.ProducerType.SYNC_LEADER_ACK;
+        final KafkaClient.ProducerType PRODUCER_TYPE = KafkaClient.ProducerType.LEADER_ACK;
 
         try (KafkaClient kafkaClient = new KafkaClient(BOOTSTRAP_SERVERS)) {
             kafkaClient.init();
+            flush(kafkaClient, GROUP_ID, TOPIC);
 
             long timestart = System.currentTimeMillis();
             long RECEIVED_MSGS = 0;
@@ -109,12 +88,13 @@ public class QndConsumerManual {
     }
 
     public void qndSyncAllAcksProducer() throws Exception {
-        System.out.println("========== QND: SyncAllAcks Producer ==========");
+        System.out.println("========== QND: AllAcks Producer ==========");
         final boolean CONSUME_FROM_BEGINNING = true;
-        final KafkaClient.ProducerType PRODUCER_TYPE = KafkaClient.ProducerType.SYNC_ALL_ACKS;
+        final KafkaClient.ProducerType PRODUCER_TYPE = KafkaClient.ProducerType.ALL_ACKS;
 
         try (KafkaClient kafkaClient = new KafkaClient(BOOTSTRAP_SERVERS)) {
             kafkaClient.init();
+            flush(kafkaClient, GROUP_ID, TOPIC);
 
             long timestart = System.currentTimeMillis();
             long RECEIVED_MSGS = 0;
@@ -143,7 +123,6 @@ public class QndConsumerManual {
     public static void main(String[] args) throws Exception {
         QndConsumerManual test = new QndConsumerManual();
 
-        test.qndAsyncProducer();
         test.qndSyncNoAckProducer();
         test.qndSyncLeaderAckProducer();
         test.qndSyncAllAcksProducer();